pushing minor changes

+++ scripts/Model_prep.py
@@ -1,8 +1,7 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-# In[2]:
-# import
+#import
 import os
 import numpy as np
 import pandas as pd
@@ -10,20 +9,31 @@ import glob
 import matplotlib.pyplot as plt
 import fiona
 import rasterio.mask
+import geopandas as gpd
 from datetime import datetime
 from rasterio.warp import reproject, Resampling
 
-# In[10]:
 
 
-# set directory
+# set up directories
 dir = "."
-Task_number = 1
+Task_number = 2
 
-# In[11]:
+if Task_number == 1:
+    modeled = os.path.join(dir, "modeled/Task1")
+else:
+    modeled = os.path.join(dir, "modeled/Task2")
 
+if Task_number == 1:
+    out_dir = os.path.join(dir, "outputs/task1/")
+else:
+    out_dir = os.path.join(dir, "outputs/task2/")
 
-lidar = glob.glob(os.path.join(dir, "lidar", "*.tif"))
+rasters_dir = os.path.join(out_dir, "rasters/")
+figs_dir = os.path.join(out_dir, "figs/")
+
+#call lidar
+lidar = glob.glob(os.path.join(dir, "lidar","*SNEX*"))
 
 # Split by underscore and pick the part that looks like a date
 parts = os.path.basename(lidar[0]).split("_")
@@ -33,7 +43,7 @@ for part in parts:
         break
 
 date_obj = datetime.strptime(date_str, "%Y%m%d")
-outfile = os.path.join(dir, "lidar", "lidar_" + date_str + ".tif")
+outfile = os.path.join(rasters_dir, "lidar_sd_"+date_str+".tif")
 
 if os.path.exists(outfile):
     print(f"Skipping existing LiDAR raster: {outfile}")
@@ -43,59 +53,59 @@ else:
         data = src.read(1, masked=True)
         profile = src.profile
         profile.update(dtype=rasterio.float32, nodata=-9999)
-        outfile = os.path.join(dir, "lidar", "lidar_" + date_str + ".tif")
-    # Replace NaNs with -9999
+
+    #Replace NaNs with -9999
     data_filled = np.where(np.isnan(data), -9999, data)
 
     # Write the new raster
     with rasterio.open(outfile, "w", **profile) as dst:
         dst.write(data_filled.astype("float32"), 1)
-
+        
     lidar_raster = outfile
 
-# In[12]:
 
-if Task_number == 1:
-    modeled = os.path.join(dir, "modeled/Task1")
+
+
+HMS_EB = glob.glob(os.path.join(modeled, "*EB_snow*.tif"))[0]
+HMS_TI = glob.glob(os.path.join(modeled, "*TI_snow*.tif"))[0]
+
+out_path = os.path.join(modeled, "HMS_EB_inches.tif")
+if os.path.exists(out_path):
+    print(f"Skipping existing LiDAR raster: {outfile}")
 else:
-    modeled = os.path.join(dir, "modeled/Task2")
-# In[13]:
+    with rasterio.open(HMS_EB) as src:
+        raster_data = src.read(1, masked=True).filled(np.nan)
+        out_raster = raster_data * 0.0254
+        
+        # Replace NaN with NoData value
+        nodata_val = -9999
+        out_raster = np.where(np.isnan(out_raster), nodata_val, out_raster)
+        
+        profile = src.profile
+        profile.update(dtype=rasterio.float32, nodata=nodata_val)
+        
+        with rasterio.open(out_path, "w", **profile) as dest:
+            dest.write(out_raster.astype("float32"), 1)
 
 
-HMS_EB = glob.glob(os.path.join(modeled, "*EB_snow_depth*.tif"))[0]
-HMS_TI = glob.glob(os.path.join(modeled, "*TI_snow_depth*.tif"))[0]
+out_path = os.path.join(modeled, "HMS_TI_inches.tif")
+if os.path.exists(out_path):
+    print(f"Skipping existing LiDAR raster: {outfile}")
+else:
+    with rasterio.open(HMS_TI) as src:
+        raster_data = src.read(1, masked=True).filled(np.nan)
+        out_raster = raster_data * 0.0254
+        
+        # Replace NaN with NoData value
+        nodata_val = -9999
+        out_raster = np.where(np.isnan(out_raster), nodata_val, out_raster)
+        
+        profile = src.profile
+        profile.update(dtype=rasterio.float32, nodata=nodata_val)
+        
+        with rasterio.open(out_path, "w", **profile) as dest:
+            dest.write(out_raster.astype("float32"), 1)
 
-with rasterio.open(HMS_EB) as src:
-    raster_data = src.read(1, masked=True).filled(np.nan)
-    out_raster = raster_data * 0.0254
-
-    # Replace NaN with NoData value
-    nodata_val = -9999
-    out_raster = np.where(np.isnan(out_raster), nodata_val, out_raster)
-
-    out_path = os.path.join(modeled, "HMS_EB_inches.tif")
-    profile = src.profile
-    profile.update(dtype=rasterio.float32, nodata=nodata_val)
-
-    with rasterio.open(out_path, "w", **profile) as dest:
-        dest.write(out_raster.astype("float32"), 1)
-
-with rasterio.open(HMS_TI) as src:
-    raster_data = src.read(1, masked=True).filled(np.nan)
-    out_raster = raster_data * 0.0254
-
-    # Replace NaN with NoData value
-    nodata_val = -9999
-    out_raster = np.where(np.isnan(out_raster), nodata_val, out_raster)
-
-    out_path = os.path.join(modeled, "HMS_TI_inches.tif")
-    profile = src.profile
-    profile.update(dtype=rasterio.float32, nodata=nodata_val)
-
-    with rasterio.open(out_path, "w", **profile) as dest:
-        dest.write(out_raster.astype("float32"), 1)
-
-# In[14]:
 
 
 rasters = {
@@ -107,40 +117,41 @@ rasters = {
 
 print(rasters)
 
-# In[15]:
-
 
 MCS = os.path.join(dir, "MCS_outline/basin_outline.shp")
 
 with fiona.open(MCS, "r") as shapefile:
     shapes = [feature["geometry"] for feature in shapefile]
 
-# In[16]:
-if Task_number == 1:
-    out_dir = os.path.join(dir, "outputs/task1/")
-else:
-    out_dir = os.path.join(dir, "outputs/task2/")
-
-rasters_dir = os.path.join(out_dir, "rasters/")
-figs_dir = os.path.join(out_dir, "figs/")
 
 stats_list = []
 
 # Loop by model
 for model, raster_list in rasters.items():
-    for raster in raster_list:
-        with rasterio.open(raster) as src:
-            out_image, out_transform = rasterio.mask.mask(src, shapes, crop=True)
-            profile = src.profile
+    out_name = f"{model}_basin_clip.tif"
+    out_path = os.path.join(rasters_dir, out_name)
 
-        # Include model in output filename
-        out_name = f"{model}_basin_clip.tif"
-        out_path = os.path.join(rasters_dir, out_name)
+    if os.path.exists(out_path):
+        print(f"Skipping existing LiDAR raster: {out_name}")
+    else:
+        for raster in raster_list:
+            with rasterio.open(raster) as src:
+                out_image, out_transform = rasterio.mask.mask(src, shapes, crop=True)
+# Copy the old profile and update it with new metadata
+                profile = src.profile
+                profile.update({
+                 "driver": "GTiff",
+                 "height": out_image.shape[1],
+                 "width": out_image.shape[2],
+                 "transform": out_transform, # <-- Use the new transform from the mask operation
+                 "nodata": -9999 # Explicitly set nodata if not already
+                 })
+            with rasterio.open(out_path, "w", **profile) as dest:
+                    dest.write(out_image)
 
-        with rasterio.open(out_path, "w", **profile) as dest:
-            dest.write(out_image)
+                
 
-        # Compute statistics
+# Compute statistics
         data = out_image  # 1 band raster, extract 2D array
         mask = (data == -9999)
         data_masked = np.ma.array(data, mask=mask)  # mask nodata
@@ -153,57 +164,74 @@ for model, raster_list in rasters.items():
             "max": data_masked.max(),
             "zeros": np.sum(data_masked == 0)
         }
-
+        
         stats_list.append(raster_stats)
+
 
 # Convert stats to a DataFrame
 stats_df = pd.DataFrame(stats_list)
 stats_csv = os.path.join(figs_dir, "basin_stats.csv")
 stats_df.to_csv(stats_csv, index=False)
 
-# In[18]:
+
 
 rasters = {
-    "HMS Energy Balance": glob.glob(os.path.join(rasters_dir, "*EB*.tif")),
-    "HMS Temperature Index": glob.glob(os.path.join(rasters_dir, "*TI*.tif")),
-    "iSnobal": glob.glob(os.path.join(rasters_dir, "*iSnobal*.tif")),
-    "SnowModel": glob.glob(os.path.join(rasters_dir, "*SnowModel*.tif")),
+    "HMS Energy Balance": glob.glob(os.path.join(rasters_dir, "*HMS_EB_basin_clip*.tif")),
+    "HMS Temperature Index": glob.glob(os.path.join(rasters_dir, "*HMS_TI_basin_clip*.tif")),
+    "iSnobal": glob.glob(os.path.join(rasters_dir, "*iSnobal_basin_clip*.tif")),
+    "SnowModel": glob.glob(os.path.join(rasters_dir, "*SnowModel_basin_clip*.tif")),
 }
 
-for model, raster_list in rasters.items():
-    print(model, len(raster_list), raster_list)
 
-# In[20]:
+#create figure of basin results
 
+lidar_shp = gpd.read_file(os.path.join(dir, "MCS_outline/MCS_outline.shp"))
 
-fig, axes = plt.subplots(2, 2, figsize=(8, 10))
+fig, axes = plt.subplots(2, 2, figsize=(8, 10), sharex=True, sharey=True)
 axes = axes.flatten()  # flatten to 1D array for easy looping
 
 for i, (model, raster_list) in enumerate(rasters.items()):
-    raster_path = raster_list[0]
-
+    raster_path = raster_list[0] 
+    
     with rasterio.open(raster_path) as src:
         data = src.read(1, masked=True)  # read first band, mask NoData
         extent = [src.bounds.left, src.bounds.right, src.bounds.bottom, src.bounds.top]
-
-    im = axes[i].imshow(data, cmap="viridis", extent=extent, vmin=0, vmax=4.0)
+    
+    im = axes[i].imshow(data, cmap="viridis",
+                        extent=extent,
+                        vmin=0,vmax=4.0 )
     axes[i].set_title(model)
-    axes[i].axis("off")
+    axes[i].grid(False)
+    lidar_shp.plot(ax=axes[i], facecolor='none', edgecolor='black', linewidth=1.5)
+
 
 # Add colorbar
-fig.colorbar(im, ax=axes, orientation="vertical", fraction=0.02, label="Snow Depth")
-fig.suptitle(f"Mores Creek Basin Snow Depth, {date_obj.strftime('%B %d, %Y')}", fontsize=16, y=0.95)
+
+#fig.colorbar(im, ax=axes, orientation="vertical", fraction=0.02, extend="max", label="Snow Depth")
+
+# --- 5. Add a single, shared colorbar ---
+fig.tight_layout(rect=[0, 0, 0.9, 0.95]) # Adjust layout to make space
+
+# Add axes for the colorbar [left, bottom, width, height]
+cbar_ax = fig.add_axes([0.92, 0.15, 0.025, 0.7])
+cbar = fig.colorbar(im, cax=cbar_ax)
+cbar.set_label('Snow Depth (m)', fontsize=14)
+
+fig.suptitle(f"Mores Creek Basin Snow Depth, {date_obj.strftime('%B %d, %Y')}", fontsize=16,  y=0.95)
 
 fig.subplots_adjust(
     wspace=0.05,
     hspace=0.06,
     right=0.88
-    # top=0.90
+    #top=0.90
 )
 
 plt.savefig(os.path.join(figs_dir, "Basin_models.png"), dpi=300, bbox_inches="tight")
 
 plt.show()
+
+
+
 
 # In[21]:
 
@@ -217,17 +245,17 @@ for model, raster_list in rasters.items():
             mask = (data == -9999)
             data_masked = np.ma.array(data, mask=mask)  # mask nodata
             flattened = data_masked.compressed()
-
-            # convert to DataFrame
+            
+        # convert to DataFrame
             df = pd.DataFrame({
-                "Model": model,  # this column will store model names
-                "value": flattened  # this column stores raster values
-            })
+            "Model": model,        # this column will store model names
+            "value": flattened     # this column stores raster values
+        })
         dfs.append(df)
 
 all_data = pd.concat(dfs, ignore_index=True)
 
-# In[22]:
+
 
 
 fig, ax = plt.subplots(figsize=(10, 6))
@@ -260,11 +288,10 @@ fig.suptitle(f"Mores Creek Snow Depth, {date_obj.strftime('%B %d, %Y')}", fontsi
 ax.set_ylabel("Snow Depth (m)")
 ax.set_xlabel("")
 
-# plt.tight_layout(rect=[0, 0, 1, 0.92])
+#plt.tight_layout(rect=[0, 0, 1, 0.92])
 plt.savefig(os.path.join(figs_dir, "basin_boxplot.png"), dpi=300, bbox_inches="tight")
 plt.show()
 
-# In[23]:
 
 
 MCS = os.path.join(dir, "MCS_outline/MCS_outline.shp")
@@ -280,7 +307,6 @@ rasters = {
     "LiDAR": [lidar_raster]
 }
 
-# In[24]:
 
 
 stats_list = []
@@ -288,46 +314,53 @@ stats_list = []
 # Loop by model
 for model, raster_list in rasters.items():
     for raster in raster_list:
-        with rasterio.open(raster) as src:
-            out_image, out_transform = rasterio.mask.mask(src, shapes, crop=True)
-            out_meta = src.meta.copy()
-
-        out_meta.update({
-            "driver": "GTiff",
-            "height": out_image.shape[1],
-            "width": out_image.shape[2],
-            "transform": out_transform,
-        })
-
-        # Include model in output filename
         out_name = f"{model}_MCS_clip.tif"
         out_path = os.path.join(rasters_dir, out_name)
+        
+        if os.path.exists(out_path):
+            print(f"skipping creation of out_name")
+            
+        else:
+            with rasterio.open(raster) as src:
+                out_image, out_transform = rasterio.mask.mask(src, shapes, crop=True)
+                out_meta = src.meta.copy()
 
-        with rasterio.open(out_path, "w", **out_meta) as dest:
-            dest.write(out_image)
+            out_meta.update({
+                "driver": "GTiff",
+                "height": out_image.shape[1],
+                "width": out_image.shape[2],
+                "transform": out_transform,
+            })
 
-        # Compute statistics
-        data = out_image[0]  # 1 band raster, extract 2D array
-        mask = (data == -9999)
-        data_masked = np.ma.array(data, mask=mask)  # mask nodata
+            # Include model in output filename
+            out_name = f"{model}_MCS_clip.tif"
+            out_path = os.path.join(rasters_dir, out_name)
 
-        raster_stats = {
-            "file": out_name,
-            "model": model,
-            "min": data_masked.min(),
-            "mean": data_masked.mean(),
-            "max": data_masked.max(),
-            "zeros": np.sum(data_masked == 0)
-        }
+            with rasterio.open(out_path, "w", **out_meta) as dest:
+                dest.write(out_image)
 
-        stats_list.append(raster_stats)
+    # Compute statistics
+            data = out_image[0]  # 1 band raster, extract 2D array
+            mask = (data == -9999)
+            data_masked = np.ma.array(data, mask=mask)  # mask nodata
+
+            raster_stats = {
+                "file": out_name,
+                "model": model,
+                "min": data_masked.min(),
+                "mean": data_masked.mean(),
+                "max": data_masked.max(),
+                "zeros": np.sum(data_masked == 0)
+            }
+            
+            stats_list.append(raster_stats)
+
 
 # Convert stats to a DataFrame
 stats_df = pd.DataFrame(stats_list)
 stats_csv = os.path.join(figs_dir, "MCS_stats.csv")
 stats_df.to_csv(stats_csv, index=False)
 
-# In[25]:
 
 
 rasters = {
@@ -338,28 +371,27 @@ rasters = {
     "LiDAR": os.path.join(rasters_dir, "LiDAR_MCS_clip.tif")
 }
 
-# In[39]:
 
-
-fig, axes = plt.subplots(2, 3, figsize=(8, 10), constrained_layout=True)
+fig, axes = plt.subplots(2, 3, figsize=(8, 10),constrained_layout=True)
 axes = axes.flatten()  # flatten to 1D array for easy looping
+
 
 axes = axes.flatten()
 
 for i, (model, raster_path) in enumerate(rasters.items()):
-    # raster_path = raster_list[0]
+    #raster_path = raster_list[0] 
     with rasterio.open(raster_path) as src:
         data = src.read(1, masked=True)  # read first band, mask NoData
         extent = [src.bounds.left, src.bounds.right, src.bounds.bottom, src.bounds.top]
-
-    im = axes[i].imshow(data, cmap="viridis", extent=extent, vmin=0, vmax=4.0)  # aspect="auto")
+    
+    im = axes[i].imshow(data, cmap="viridis", extent=extent, vmin=0,vmax=4.0) #aspect="auto")
     axes[i].set_title(model)
     axes[i].axis("off")
 
 axes[-1].axis("off")
 # Add colorbar
 fig.colorbar(im, ax=axes, orientation="vertical", fraction=0.02, label="Snow Depth")
-fig.suptitle(f"Mores Creek Summit Snow Depth, {date_obj.strftime('%B %d, %Y')}", fontsize=14, y=0.95)
+fig.suptitle(f"Mores Creek Summit Snow Depth, {date_obj.strftime('%B %d, %Y')}", fontsize=14,  y=0.95)
 
 # fig.subplots_adjust(
 #     #wspace=.01,
@@ -372,29 +404,28 @@ plt.savefig(os.path.join(figs_dir, "MCS_models.png"), dpi=300, bbox_inches="tigh
 
 plt.show()
 
-# In[41]:
+
 
 
 dfs = []
 
 for model, raster in rasters.items():
-    # for raster in raster_list:
-    with rasterio.open(raster) as src:
-        data = src.read(1, masked=True)
-        mask = (data == -9999)
-        data_masked = np.ma.array(data, mask=mask)  # mask nodata
-        flattened = data_masked.compressed()
-
+    #for raster in raster_list:
+        with rasterio.open(raster) as src:
+            data = src.read(1, masked=True)
+            mask = (data == -9999)
+            data_masked = np.ma.array(data, mask=mask)  # mask nodata
+            flattened = data_masked.compressed()
+            
         # convert to DataFrame
-        df = pd.DataFrame({
-            "Model": model,  # this column will store model names
-            "value": flattened  # this column stores raster values
+            df = pd.DataFrame({
+            "Model": model,        # this column will store model names
+            "value": flattened     # this column stores raster values
         })
-    dfs.append(df)
+        dfs.append(df)
 
 all_data = pd.concat(dfs, ignore_index=True)
 
-# In[42]:
 
 
 fig, ax = plt.subplots(figsize=(10, 6))
@@ -427,24 +458,21 @@ fig.suptitle(f"Mores Creek Summit Snow Depth, {date_obj.strftime('%B %d, %Y')}",
 ax.set_ylabel("Snow Depth (m)")
 ax.set_xlabel("")
 
-# plt.tight_layout(rect=[0, 0, 1, 0.92])
+#plt.tight_layout(rect=[0, 0, 1, 0.92])
 plt.savefig(os.path.join(figs_dir, "MCS_boxplot.png"), dpi=300, bbox_inches="tight")
 plt.show()
 
-# In[43]:
 
 
 del rasters["LiDAR"]
 lidar = os.path.join(rasters_dir, "LiDAR_MCS_clip.tif")
 
-# In[404]:
+
 with rasterio.open(lidar) as src:
     lidar_data = src.read(1, masked=True)
-    profile = src.profile  # for writing outputs
+    profile = src.profile                 # for writing outputs
     lidar_crs = src.crs
     lidar_transform = src.transform  # True where LiDAR is masked
-
-# In[410]:
 
 
 for model, raster_path in rasters.items():
@@ -482,14 +510,17 @@ for model, raster_path in rasters.items():
 
         # Write difference raster
         out_profile = profile.copy()
-        out_profile.update(dtype=rasterio.float32, compress="lzw")
+        out_profile.update(dtype=rasterio.float32, compress="lzw", nodata = -9999)
 
         # Write difference raster
-        out_path = os.path.join(rasters_dir, f"{model.replace(' ', '_')}_lidar_diff.tif")
-        with rasterio.open(out_path, "w", **out_profile) as dst:
+        out_path_diff = os.path.join(rasters_dir, f"{model.replace(' ', '_')}_lidar_diff.tif")
+        out_path_resample = os.path.join(rasters_dir, f"{model.replace(' ', '_')}_lidar_resample.tif")
+        
+        with rasterio.open(out_path_resample, "w", **out_profile) as dst:
+            dst.write(model_masked.filled(np.nan).astype(np.float32), 1)
+        with rasterio.open(out_path_diff, "w", **out_profile) as dst:
             dst.write(diff_data.filled(np.nan).astype(np.float32), 1)
 
-# In[ ]:
